Adapted sales views to user abstract classes
