+++ sales/urls.py
@@ -3,12 +3,13 @@ from . import views
 
 urlpatterns=[
     path('', views.invoicelist, name='sales-list'),
-    path('create/', views.InvoiceCreateView.as_view(), name='sales-create'),
-    path('<int:pk>/edit/', views.InvoiceUpdateView.as_view(), name='sales-edit'),
-    path('<int:pk>/delete/', views.InvoiceDeleteView.as_view(), name='sales-delete'),
-    path('<int:pk>/view/', views.InvoiceDetailView.as_view(), name='sales-view'),
+    path('create/', views.SalesInvoiceCreateView.as_view(), name='sales-create'),
+    path('<int:pk>/edit/', views.SalesInvoiceUpdateView.as_view(), name='sales-edit'),
+    path('<int:pk>/view/', views.SalesInvoiceDetailView.as_view(), name='sales-view'),
+    path('<int:pk>/delete/', views.SalesInvoiceDeleteView.as_view(), name='sales-delete'),
     path('<int:pk>/entry/add/',
-         views.InvoiceEntryCreateView.as_view(), name='sales-entry-add'),
+         views.SalesInvoiceEntryCreateView.as_view(), name='sales-entry-add'),
     path('entry/<int:pk>/delete/',
-         views.InvoiceEntryDeleteView.as_view(), name='sales-entry-delete'),
+         views.SalesInvoiceEntryDeleteView.as_view(), name='sales-entry-delete'),
+
 ]
