+++ sales/views.py
@@ -1,12 +1,18 @@
-from django.shortcuts import render, redirect, get_object_or_404
+from django.shortcuts import render
 from django.urls import reverse_lazy
-from django.views.generic import CreateView, UpdateView, DetailView, DeleteView
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.core.paginator import Paginator
 from .models import SalesInvoice, SalesInvoiceEntry
 from .filters import SalesInvoiceFilter
 from .forms import SalesInvoiceForm, SalesInvoiceEntryForm
+
+from products.abstract_views import (
+    InvoiceCreateView, 
+    InvoiceDetailView, 
+    InvoiceUpdateView, 
+    InvoiceDeleteView,
+    InvoiceEntryCreateView,
+    InvoiceEntryDeleteView)
 
 # Create your views here.
 @login_required
@@ -23,85 +29,27 @@ def invoicelist(request):
     return render(request, "sales/salesinvoice_list.html", context=context)
 
 
-class InvoiceCreateView(LoginRequiredMixin, CreateView):
+class SalesInvoiceCreateView(InvoiceCreateView):
     model = SalesInvoice
     form_class = SalesInvoiceForm
 
-    def form_valid(self, form):
-        form.instance.organization = self.request.user.organization
-        return super().form_valid(form)
-
-    def get_success_url(self):
-        return self.object.get_absolute_url()
-
-class InvoiceUpdateView(LoginRequiredMixin, UserPassesTestMixin, UpdateView):
+class SalesInvoiceUpdateView(InvoiceUpdateView):
     model = SalesInvoice
     form_class = SalesInvoiceForm
-
-    def form_valid(self, form):
-        form.instance.organization = self.request.user.organization
-        return super().form_valid(form)
-    
-    def test_func(self):
-        invoice = self.get_object()
-        return invoice.organization == self.request.user.organization and not invoice.finalized #Don't let the user edit if invoice is finalized
    
-class InvoiceDetailView(LoginRequiredMixin, UserPassesTestMixin, DetailView):
+class SalesInvoiceDetailView(InvoiceDetailView):
     model = SalesInvoice
+    invoice_finalize_redirect_name = 'sales-view'
 
-    def post(self, request, pk, *args, **kwargs):
-        #Set invoice as finalized
-        invoice = self.get_object()
-        #Reduce inventory
-        invoice.finalize(action=invoice.SALE)
-        return redirect("sales-view", pk=pk)
-
-    def test_func(self):
-        invoice = self.get_object()
-        return invoice.organization == self.request.user.organization
-
-class InvoiceDeleteView(LoginRequiredMixin, UserPassesTestMixin, DeleteView):
+class SalesInvoiceDeleteView(InvoiceDeleteView):
     model = SalesInvoice
     success_url = reverse_lazy('sales-list')
 
-    def test_func(self):
-        invoice = self.get_object()
-        # Don't let the user edit if invoice is finalized
-        return invoice.organization == self.request.user.organization
-
-class InvoiceEntryCreateView(LoginRequiredMixin, UserPassesTestMixin, CreateView):
+class SalesInvoiceEntryCreateView(InvoiceEntryCreateView):
     model = SalesInvoiceEntry
+    parent_model = SalesInvoice
     form_class = SalesInvoiceEntryForm
-    
-    def form_valid(self, form):
-        self.invoice = get_object_or_404(
-            SalesInvoice, pk=self.kwargs.get('pk'))
-        form.instance.invoice = self.invoice
-        return super().form_valid(form)
-    
-    def get_success_url(self):
-        return reverse_lazy('sales-entry-add', args=[self.invoice.pk])
-
-    def get_context_data(self, *args, **kwargs):
-        context = super().get_context_data(*args, **kwargs)
         
-        context['pk'] = self.invoice.pk
-        context['added_entries'] = self.invoice.entries.all()
-        return context
-    
-    def test_func(self):
-        self.invoice = get_object_or_404(SalesInvoice, pk=self.kwargs.get("pk"))
-        # Don't let the user edit if invoice is finalized
-        return self.invoice.organization == self.request.user.organization and not self.invoice.finalized
-        
-class InvoiceEntryDeleteView(LoginRequiredMixin, UserPassesTestMixin, DeleteView):
 
+class SalesInvoiceEntryDeleteView(InvoiceEntryDeleteView):
     model = SalesInvoiceEntry
-
-    def get_success_url(self):
-        return self.invoice.get_absolute_url()
-
-    def test_func(self):
-        self.invoice = self.get_object().invoice
-        # Don't let the user edit if invoice is finalized
-        return self.invoice.organization == self.request.user.organization and not self.invoice.finalized
